LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-initial-focus-display-animation-expected.txt



FAIL Option initial focus should still work when a display animation is present. assert_equals: expected Element node <option>option</option> but got Element node <select>
  <option>option</option>
</select>
